requirements.txt
# Core Data Science Libraries
pandas==2.0.3
numpy==1.24.3
scipy==1.11.2

# Machine Learning
scikit-learn==1.3.0
xgboost==1.7.6
imbalanced-learn==0.11.0

# Visualization
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.17.0

# Jupyter
jupyter==1.0.0
ipykernel==6.25.2
notebook==7.0.3

# API and Web
fastapi==0.103.0
uvicorn==0.23.2
pydantic==2.3.0
python-multipart==0.0.6

# Configuration
pyyaml==6.0.1
python-dotenv==1.0.0

# Testing
pytest==7.4.0
pytest-cov==4.1.0

# Model Interpretation
shap==0.42.1

# Data Validation
great-expectations==0.17.23

# Utilities
joblib==1.3.2
tqdm==4.66.1

# Kaggle API
kaggle==1.5.16

# Monitoring
evidently==0.4.5

# Logging
colorlog==6.7.0
